+++ src/Vector2.h
@@ -3,7 +3,7 @@
 
 class Vector2 : public Vector {
 public:
-	Vector2(double x, double y);
+	Vector2(float x, float y);
 	Vector2(const Vector& vec);
 	Vector2();
 	// @return (x, y)
